Added Automatic Collider Creator
Changes:
* New createCollider method inside of the GameObject class allows for 
the easy creation of snug colliders on a GameObject.
* Updated devList to reflect recent changes.
* Removed colliderObject member from gameObjectInfo struct.

Notes:
* To create a collider, all that needs to happen is after GameObject 
creation, get a reference to the GameObject you want to add a collider 
to and simply call the createCollider method. The single argument to 
this method is the programID of the shader used for rendering wireframe 
boxes (can just make -1 if you don't want the collider to be rendered).