Added Anti-Aliasing, Fixed Race Condition
Changes:
* Added anti-aliasing by default for smoother visuals.
* Fixed a race condition that caused the Dracula object and other 
objects to randomly flicker across the screen as they were moved. 
Movement is now much smoother because of this fix.

Notes:
* The bug that caused race conditions was due to being able to update 
individual objects between draw calls, which could make some objects 
move sooner than others. This was fixed by locking the entire scene when 
drawing it, and unlocking when finished. Because of this new method, the 
main loop responsible for updating the game scene must delay at the end 
to allow the other threads to execute. 
* Locking individual GameObjects using the lockObject method is still 
supported at the moment, but should not be used any more as it will be 
removed from future versions. Currently information for camera and text 
objects are protected using their lockObject methods, but their code 
areas should be re-written to move away from relying on these locks for 
protection.