Remove GTest as Build Requirement
Move GTest out of being a build requirement. We may want to split this off into its own project in the future. GTest shouldn't be required to be installed in order to build. This will remove that depency and just not create the GTest files if CMake doesn't find GTest.